Renamed `NonAssociationsDetails` to `NonAssociationDetails` singular
Not sure yet if all non-associations things should be have a plural
'associations' in the name. Prison API's endpoints use the singular but
then uses the plural in the response for the list of non-associations.

At this stage I think:
- all things like resources/services/etc...should use the plural as they
  handle multiple non-associations
- a single non-association should use the singular
- a list or collection of non-associations should use the plural

This is likely to cause some confusion as it's easy to miss the final 's'.
However always use the plural or the singular seems wrong, for example
I wouldn't call a DB row/record for a non-association between two prisoners
a non-associations*S*.