Add PrisonApiService to make requests to Prison API
I tried to avoid using `suspend fun` functions by using `runBlocking {}`
but it doesn't seem to work as it starts to complain about Authentication
key not being present in the context.

Someone with a better knowledge of Spring Boot and its ecosystem/packages
can probably get something working [without using `suspend`] but for now
at least this works fine and it's very similar to what we do in the
Incentives API.

NOTE: The DTO still contains a single field so not all data from Prison API
is returned at the moment.