NON-10: Added `PrisonApiService` to make requests to Prison API (#7)
* NON-10: WebClient configuration to talk with APIs

Copied/adapted from Incentives API

* Moved non-associations details DTO class under dto.prisonapi

We may have different DTO for our service and it's good to know this is
the DTO for the format returned by the Prison API in case we tweak it
for whatever reasons

* Renamed `NonAssociationsDetails` to `NonAssociationDetails` singular

Not sure yet if all non-associations things should be have a plural
'associations' in the name. Prison API's endpoints use the singular but
then uses the plural in the response for the list of non-associations.

At this stage I think:
- all things like resources/services/etc...should use the plural as they
  handle multiple non-associations
- a single non-association should use the singular
- a list or collection of non-associations should use the plural

This is likely to cause some confusion as it's easy to miss the final 's'.
However always use the plural or the singular seems wrong, for example
I wouldn't call a DB row/record for a non-association between two prisoners
a non-associations*S*.

* Legacy Prison APIs return NA details, not a list, rename accordingly

* Added `/legacy/` prefix to Prison API proxy endpoints

Make it clear these endpoints are not new to the Non Associations service.
The main point is that we want to clearly distinguish between the old
formats and the new ones (in case we need to make changes to the structure
of the requests/responses for whatever reason)

* Added skeleton of NA stubs for Prison API mock

And removed a bunch of stubs for endpoints that we're probably not going
to use (in the short term anyway). We can always add more later on.

NOTE: The stubbed response is incomplete.

* Add PrisonApiService to make requests to Prison API

I tried to avoid using `suspend fun` functions by using `runBlocking {}`
but it doesn't seem to work as it starts to complain about Authentication
key not being present in the context.

Someone with a better knowledge of Spring Boot and its ecosystem/packages
can probably get something working [without using `suspend`] but for now
at least this works fine and it's very similar to what we do in the
Incentives API.

NOTE: The DTO still contains a single field so not all data from Prison API
is returned at the moment.