NON-10: 2nd endpoint, better docs, catch 404s, etc... (#11)
* Added Swagger documentation for DTOs fields

* Added `jsonString()` method to base test class

A bit nicer and clear than using Jackson's ObjectMapper directly

* Tests for endpoint to get non-associations details by prisoner number

* Renamed Prison API stub methods for clarity

Makes sense for the method that stubs the endpoint to get non-associations
by prisoner number to use the prisoner number in the expected response
(to be sure it's consistent). I've renamed the other stub method for
consistency.

* Read non-associations-api OAuth client creds from k8s Secret

* Added endpoint to get non-association details by prisoner number

* Deprecate endpoint to get details by booking Id

When it comes to non-associations the person (e.g. the prisoner numbers)
is what matter rather than a stretch of time in prison (booking Ids).

Clients should get non-associations by prisoner number. They probably
have this information already at hand so it should be trivial to change
any existing code to get non-association details by prisoner number.

* NON-10: Handle 404 Not Found gracefully

Respond with a 404 instead than with a 500 Internal Server Error.

* Typo for swagger doc for prisoner number endpoint

* Removed endpoint to get non-associations by Booking ID

Clients should have the Prisoner number available and booking IDs
can change when prisoners get out/in prison which makes them not ideal
when dealing with non-associations.

* Lint