NON-10: Optional assigned living unit fields
The assigned living unit fields could be null and in general a non-association
could be with someone who left prison (but may return, and that's why
it's imporant to still retain the non-association)

I've also update the Swagger documentation for the fields accordingly.

NOTE: While I was getting an exception when Spring Boot/Jackson for the
missing mandatory `String` field, it didn't complain about the missing
`Long` field and implicitly converted it to `0`.
We should distinguish between `0`s and `null`s are these are different
concept, one signifify the absence of information (e.g. there is no
assigned unit) while the other imply knowledge (e.g. the assigned unit
is XYZ). We don't want `0` to be possibly interpreted as an ID for
a real living unit as that would be wrong.

That's why the `Long?` optional fields have a default value of `null` now.