Renamed dto.prisonapi classes to avoid swagger clash with new dto class
The swagger documentation for the response of legacy endpoint was
incorrectly displaying the (WIP) format of the new endpoint.
So I've prefixed the Prison API dto classes with `Legacy`.
Aside from the swagger docs now displaying correctly this shouldn't
make any other difference and it should help distinguish these classes
from the new ones.

I've also simplified the PrisonApiService call with
`.bodyToMono(LegacyNonAssociationDetails::class.java)` as it compiles
and it seems to work the same, but it reads better.