NON-77: Prevent closing a non-association without providing details
Added DB-level checks that prevent closing a non-association without
providing details about the reason/who closed it and when.

If a non-association is closed these details have to be present.
If a non-association is open (is_closed=false) these details have to be
null. This is subtle but it's to prevent possible stale information if
a non-association is closed-reopened-closed: A subsequent close would
likely have different reasons for closing and a different time of closure
for certain.

I've added some tests related to closing a non-association and a test
for the `whenUpdated` field. This work but when running tests the
`TestTransaction` has to be manually committed/ended to avoid surprising
results caused by separate updates not being committed.