NON-77: Use enums instead of strings in JPA entity
For non-association reason and restriction type.
This is higher level and nicer than using generic `String` values - which
can contain anything.

As a result, the values stored are no longer the codes used by
NOMIS/Prison API but the enum variants names (which are longer hence the
increased column length). However this is nicer and less obscure (`PERPETRATOR`
would be clear to anyone while `PER` could be confusing).