NON-77: Non-association starter model (#27)
* NON-77: Add non-association DB table/entity/repository

It's a starting point and we may have to tweak this.

* NON-77: Added authorised by and renamed restriction code field...

...to `non_association` DB table/entity

- The value stored in the DB would be a code for the restriction type
  (e.g. `CELL` or `WING`) so renamed to avoid confusion
- Non-associations in NOMIS/Prison API have the person who authorised
  the non-association. This is currently optional **free test** there so
  I've retained the same flexibility at the moment.
- we may want to add a field to record the user ID of the person who
  created the non-association possibly alongside the `authorised_by` field
  but this will need some thinking and coordination depending how sync
  will work

* NON-77: Prevent closing a non-association without providing details

Added DB-level checks that prevent closing a non-association without
providing details about the reason/who closed it and when.

If a non-association is closed these details have to be present.
If a non-association is open (is_closed=false) these details have to be
null. This is subtle but it's to prevent possible stale information if
a non-association is closed-reopened-closed: A subsequent close would
likely have different reasons for closing and a different time of closure
for certain.

I've added some tests related to closing a non-association and a test
for the `whenUpdated` field. This work but when running tests the
`TestTransaction` has to be manually committed/ended to avoid surprising
results caused by separate updates not being committed.

* NON-77: Added NonAssociationReason enum and dropped DB table

Use a DB table was mainly to check Spring Boot/JPA wiring and was overkill
given the small number of values and how infrequently these would change.

* NON-77: Added NonAssociationRestrictionType enum and dropped DB table

Use a DB table was mainly to check Spring Boot/JPA wiring and was overkill
given the small number of values and how infrequently these would change.

NOTE: These values are taken from NOMIS/Prison API `references_codes`
DB table and we may want to tweak them (although keeping compatibility
with the existing codes may prevent confusion or even problems)

* NON-77: Add more DB restrictions

- prisoners numbers have to be present
- and different (cannot non-associate someoen with themselves)
- reason codes have to be present
- restriction type code has to be present

* NON-77: Added indexes to non_association table

to the prisoner numbers columns and to the incident report number which
is something worth indexing.

* NON-77: Use enums instead of strings in JPA entity

For non-association reason and restriction type.
This is higher level and nicer than using generic `String` values - which
can contain anything.

As a result, the values stored are no longer the codes used by
NOMIS/Prison API but the enum variants names (which are longer hence the
increased column length). However this is nicer and less obscure (`PERPETRATOR`
would be clear to anyone while `PER` could be confusing).

* NON-77: Add method to close non-association

as suggested by Michael we can incapsulate this operation inside a JPA
entity method.