NON-87: Added method to get non-association by ID
Note that this method returns an optional non-association as this may not
exist.

Also note that we don't throw an exception in this context, as we don't
have enough information to make that call. For example a request handler
would want to throw a "not found" exception but in other places we may
just ignore this (and handle exceptions just to determine if something
exists is awkward)