NON-88: Create a non-association checks prisoner numbers are valid
If not it responds 404 Not Found rather than creating records with invalid
values.

OffenderSearch API is used to get the information about the given prisoners
(numbers). This now checks if all the searched prisoners were found
otherwise throws an exception. Makes sense to have this in the Offender
Search method as I'd imagine we always want to be sure all the searched
prisoners were found.

Offender Search now returns a map rather than a list, this is probably
going to be more convinient when hydrating non-associations with the
prisoners details.