NON-88: Get non-associations for a prisoner (#35)
* NON-88: Added OffenderSearch service to search for prisoners

This is needed for a number of reasons, for example validation (e.g.
is this prisonerNumber valid?) but also to only return non-associations
in the same prison rather than across the estate (by default)

* NON-88: Create a non-association checks prisoner numbers are valid

If not it responds 404 Not Found rather than creating records with invalid
values.

OffenderSearch API is used to get the information about the given prisoners
(numbers). This now checks if all the searched prisoners were found
otherwise throws an exception. Makes sense to have this in the Offender
Search method as I'd imagine we always want to be sure all the searched
prisoners were found.

Offender Search now returns a map rather than a list, this is probably
going to be more convinient when hydrating non-associations with the
prisoners details.

* NON-88: Updated DTO returned when getting a prisoner's non-associations

* docker-compose-local: Moved services to docker-compose

We usually run the API locally against dev environment, so the
docker-compose-local.yml only needs the DB and the localstack containers.

Rest of services were moved into the docker-compose.yml which is (presumably)
used to run the whole service and its dependencies.

* Renamed method to get non-associations details from Prison API

I want to keep it separate for now and name them more explicitly.

* Exposed whenCreated field in NonAssociation DTO

* NON-88: Better tests for get non-associations for a prisoner

plus using the offender search test data

NOTE: Some of the classes/methods used may be in a separate commit

* NON-88: Add way to get a prisoner non-associations

regardless of whether their prisonerNumber is stored on the lhs or rhs.

Formatting of these non-associations into a `PrisonerNonAssociations`
data class is done separately by a method on `List<NonAssociationJPA>`
which formats these into the API response structure.

* NON-88: Format a prisoner non-associations into the DTO response

Method that converts a list of `NonAssociationJPA` into an instance of
`NonAssociationDetails`.

The method "flips" first/second prisoner in the non-association record
so that the "main" prisoner is shown before the other prisoner.

* Renamed OffenderSearchPrisoner file

* add missing imports in resource test

* NON-88: Fix flaky test in CI to do with `LocalDateTime` serialisation

In CI this test was failing because of the `whenCreated` field was not
*exactly* the same. I'm using a JSON string directly now and removed the
`whenCreated` field from the test as we don't really care about the
specific datetime in this context (and it's mandatory so enforced by the
type)

* NON-88: Create records directly in resource test

Using the repository instead of using the API

* Renamed list options for getting a prisoner non-associations

* Use API creds by default to make requests to Offender Search API

As discussed in PR comments and offline we don't want the user to need to
have too many permissions and gathering prisoners information is internal
to the API.