NON-88: Ability to sort a prisoner non-associations (#37)
* Ability to control order of returned prisoner's non-associations

I may be able to simplify the comparators logic, but this should work
(needs some more testing)

* Gracefully handle invalid arguments for enums

Respond with a 400 Bad Request with a useful message (e.g. with the
valid values) rather than with a 500 Internal Server Error

* Test for prisoner non-associations sorting

* Simplified comparator logic (no need for the Comparator subclasses)

* Ability to sort prisoner non-associations by first name (other prisoner)

This could be useful. More sorting options could be added easily as
necessary.

* Ability to sort by prisoner number