Add a note in `getPrisonerNonAssociations()`
The way prisoner numbers are collected is slightly different to account
for prisoners without any non-associations. We want to be sure the
information about the "key" prisoner are always retrieved from Offender
Search API, even when that's the case.

This is important because the `toPrisonerNonAssociations()` method called
on the list of filtered non-associations always need the information about
this "key" prisoners to construct the response