NON-347: Show only open non-associations by default
I know this is just a spike and a "dummy" test report but from what I can
tell from [this spreadsheet] existing NOMIS reports are concerned with
"current, active Non Association", aka open non-associations so it makes
sense to replicate that in this test report.

It is also an excuse to provide an example of default filter.

[this spreadsheet]: https://docs.google.com/spreadsheets/d/11kljKuJ7nzOSaNnEzh-DmS1oEAZPy7y3o2GwKC9tO04/edit#gid=0