NON-358: SAR endpoint responds `204 No Content` when no information matching
The endpoint now responds `204 No Content` when no non-associations for
the given prisoner were found or when none of them were within the given
date range.

**NOTE**: Offender Search API could potentially not have information on some
of the prisoners involved in the non-associations found. This is an edge
case but without these information (names, etc...) we cannot construct
the response correctly so the endpoint will continue to respond
`404 Not Found` in these cases.

This is also better than responding `204 No Content` and misleading the
client into believing there were no information on the given subject.
It shouldn't happen in theory anyway.