NON-358: Removed unnecessary try/catch in SAR endpoint
The repository method used to get non-associations does not throw a
`NotFound` exception (just an empty list when none are found).
We already check this condition and respond with a `204 No Content` as it
should.

The `try/catch` was probably due to the test not mocking the request to
Offender Search API, I've tweaked the test and it passes even without
that `try/catch`.