NON-358: SAR endpoint responds `204 No Content` when no information matching (#255)
* NON-358: SAR endpoint responds `204 No Content` when no information matching

The endpoint now responds `204 No Content` when no non-associations for
the given prisoner were found or when none of them were within the given
date range.

**NOTE**: Offender Search API could potentially not have information on some
of the prisoners involved in the non-associations found. This is an edge
case but without these information (names, etc...) we cannot construct
the response correctly so the endpoint will continue to respond
`404 Not Found` in these cases.

This is also better than responding `204 No Content` and misleading the
client into believing there were no information on the given subject.
It shouldn't happen in theory anyway.

* NON-358: I, for one, welcome our linter overlords

* NON-358: Removed unnecessary try/catch in SAR endpoint

The repository method used to get non-associations does not throw a
`NotFound` exception (just an empty list when none are found).
We already check this condition and respond with a `204 No Content` as it
should.

The `try/catch` was probably due to the test not mocking the request to
Offender Search API, I've tweaked the test and it passes even without
that `try/catch`.