Use single principal as token cache key (#486)
* Use single principal as token cache key

Copied/adapted from: https://github.com/ministryofjustice/hmpps-manage-users-api/pull/319/files

The `ClientCachingOAuth2AuthorizedClientService` has since been updated
to work with newer versions of Spring Boot.

* linting

* Load OAuth2ClientConfiguration after ClientRegistrationRepository beans

A lot of tests were failing because of no registrations being available
even if they should, a bit of digging and it may be down to the order
of initializations of the beans.