IR-1354: Updated Node.js 24 dependencies (#543)
* Updated types definitions to use Node.js 24 types

* Run npm run format

* Prevent npm scripts to run + saner npm config

* Use Node.js 24 and upgraded HMPPS rest-client

The version of Node.js in the nvmrc is mainly used
to inform local development version.

The `@ministryofjustice/hmpps-rest-client` package
has been upgrade to support Node.js 24, this is useful
to allow to be stricter on the UI with the engine
version (potentially, assuming all other dependencies
support Node.js 24 as well).

NOTE: I had to tweak some of the tests assertions
on the amount of logged messages. Presumably this
is to do with the newer version of the rest-client
being a bit chattier, it should break anything in
theory.